requirements.txt
Django==3.2
djangorestframework
django-cors-headers
django-filter
django-environ
psycopg2-binary
psycopg2
